bugfix pluginPaths configuration missing in 1.3
again, i see the pluginPaths configuration is missing, and this time I've fixed it by defining a reasonable default array instead of an empty array.
went ahead and assigned to a configuration so we are not re-defaulting for multiple requests.

related to previous commit 5661f98e3ea4eb991ffcd615bcc46c46a46f8932
https://github.com/CakeDC/migrations/commit/5661f98e3ea4eb991ffcd615bcc46c46a46f8932