Fix failing test - testRun
The test was failing due to the assertion on line 238, where it expected 'reset' => true to exist in the array - but there's no situation where that key/value is added, or needs to be added, in the code itself. Reset works, and asserting that 'version' => 0 is sufficient to test this.

Signed-off-by: Joshua Paling <joshua.paling@gmail.com>