Fixes Schema/app.php issue.
The SchemaShell has two possible arguments: `--name` and `--file`. If you only specify `--name` (as this PR does), then the `--file` param is automatically set to match. This means you''ll end up with a `Config/Schema/app.php` file instead of `Config/Schema/schema.php as expected.

This PR should be safe to apply once cakephp/cakephp#4230 is merged and available in a production release of the core.