Remove testTruncateLongIndexKey
It tested if a SQL error was thrown when creating a unique index 1024 bytes long, but modern MySQL 5.7 installations with InnoDB tables allow indexes as long as 3072 bytes, msSQL 900 bytes and Postgres seems to have only a 1GB limit. This unit test was testing the database engine capabilities, not the plugin code.